feat: move test support into own module
